requirements.txt
Django==1.7.1
dj-database-url==0.3.0
django-uuidfield
django-push-notifications
gunicorn==18.0
psycopg2==2.6.1
requests==2.8.1
nltk==3.0.5
numpy==1.10.1
paho-mqtt
